long long to int64_t

+++ builtin_op_importers.cpp
@@ -1687,8 +1687,8 @@ DEFINE_BUILTIN_OP_IMPORTER(Slice) {
   // (TRT doesn't support batch dim slicing)
   ASSERT(attrs.count("axes"), ErrorCode::kUNSUPPORTED_NODE);
   auto axes = attrs.get<std::vector<int>>("axes");
-  auto starts = attrs.get<std::vector<long long>>("starts");
-  auto ends = attrs.get<std::vector<long long>>("ends");
+  auto starts = attrs.get<std::vector<int64_t>>("starts");
+  auto ends = attrs.get<std::vector<int64_t>>("ends");
   int H_idx = -1, W_idx = -1;
   // This will be modified to contain expanded dims metadata,
   // if needed (e.g. input tensor is 2D)
